🚧 Refactor monorepo structure
