+++ pkg/shared/resources.go
