+++ pkg/shared/type_character.go
