+++ pkg/shared/type_character_attributes.go
