+++ pkg/shared/type_character_corporation.go
