+++ pkg/shared/type_character_skill.go
