+++ pkg/shared/type_character_skillqueue.go
