+++ pkg/skillpath/calculator.go
