+++ pkg/updater/character_attributes.go
