+++ pkg/updater/character_info.go
