+++ pkg/updater/character_skills.go
