+++ pkg/updater/character_updater.go
